\name{alternatives}
\alias{alternatives}
\title{alternatives}
\description{Make a pattern that matches one of the specified alternatives.}
\usage{alternatives(...)}
\arguments{
  \item{\dots}{Each argument is a different alternative pattern.}
}

\value{Pattern list.}

\author{Toby Dylan Hocking}





