\name{check_df_names}
\alias{check_df_names}
\title{check df names}
\description{Check that first argument is a data frame and then call
\code{\link{check_names}} on its names.}
\usage{check_df_names(...)}
\arguments{
  \item{\dots}{data frame, regex pattern args.}
}



\author{Toby Dylan Hocking}





