\name{check_names}
\alias{check_names}
\title{check names}
\description{Check that \code{subject} is a vector of unique names and then call
\code{\link{capture_first_vec}}.}
\usage{check_names(subject, 
    var.args)}
\arguments{
  \item{subject}{
}
  \item{var.args}{
}
}



\author{Toby Dylan Hocking}





