\name{melt_list}
\alias{melt_list}
\title{melt list}
\description{Compute a list of arguments to pass to
\code{\link[data.table]{melt.data.table}}.}
\usage{melt_list(measure.fun, 
    dot.args, ...)}
\arguments{
  \item{measure.fun}{
}
  \item{dot.args}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking}





