\name{stop_for_capture_same_as_id}
\alias{stop_for_capture_same_as_id}
\title{stop for capture same as id}
\description{Error if capture names same as \code{id.vars}.}
\usage{stop_for_capture_same_as_id(capture.vars, 
    id.vars)}
\arguments{
  \item{capture.vars}{
}
  \item{id.vars}{
}
}



\author{Toby Dylan Hocking}





