\name{stop_for_engine}
\alias{stop_for_engine}
\title{stop for engine}
\description{Stop if specified \code{engine} is not available.}
\usage{stop_for_engine(engine)}
\arguments{
  \item{engine}{character string: PCRE, RE2, or ICU.}
}

\value{character string.}

\author{Toby Hocking <toby.hocking@r-project.org> [aut, cre]}





