\name{try_or_stop_print_pattern}
\alias{try_or_stop_print_pattern}
\title{try or stop print pattern}
\description{Try to run a capture function. If it fails we wrap the error
message with a more informative message that also includes the
generated pattern.}
\usage{try_or_stop_print_pattern(expr, pat, engine)}
\arguments{
  \item{expr}{
}
  \item{pat}{
}
  \item{engine}{
}
}



\author{Toby Dylan Hocking}





