\name{check_names}
\alias{check_names}
\title{check names}
\description{Check that \code{subject} is a vector of unique names and then call
\code{\link{capture_first_vec}}.}
\usage{check_names(subject, 
    var.args)}
\arguments{
  \item{subject}{character vector, data frame column names.}
  \item{var.args}{regex pattern list.}
}



\author{Toby Dylan Hocking}





