\name{only_captures}
\alias{only_captures}
\title{only captures}
\description{Extract capture \code{\link{group}} columns from \code{match.mat} and assign optional
groups to "".}
\usage{only_captures(match.mat, 
    stop.fun)}
\arguments{
  \item{match.mat}{
}
  \item{stop.fun}{
}
}



\author{Toby Dylan Hocking}





