\name{subject_var_args}
\alias{subject_var_args}
\title{subject var args}
\description{Parse the complete argument list including subject.}
\usage{subject_var_args(...)}
\arguments{
  \item{\dots}{subject, regex/conversion.}
}

\value{Result of \code{\link{var_args_list}} plus subject.}

\author{Toby Dylan Hocking}





